+++ tests/backends/models.py
@@ -1,6 +1,6 @@
 from django.db import models
 
-from clickhouse_backend.models import ClickhouseModel
+from clickhouse_backend.models import ClickhouseModel, MergeTree
 
 
 class Square(ClickhouseModel):
@@ -46,14 +46,6 @@ class VeryLongModelNameZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZ(ClickhouseMode
 
 class Tag(ClickhouseModel):
     name = models.CharField(max_length=30)
-
-
-class Post(ClickhouseModel):
-    name = models.CharField(max_length=30)
-    text = models.TextField()
-
-    class Meta:
-        db_table = "CaseSensitive_Post"
 
 
 class Reporter(ClickhouseModel):
@@ -150,3 +142,13 @@ class SQLKeywordsModel(ClickhouseModel):
 
     class Meta:
         db_table = "order"
+
+
+class EngineWithSettings(ClickhouseModel):
+    class Meta:
+        engine = MergeTree(
+            order_by=(),
+            index_granularity=1024,
+            index_granularity_bytes=1 << 20,
+            enable_mixed_granularity_parts=1,
+        )